+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to **Dangl.Data.Shared** are documented here.
 
 ## v1.1.0:
 - The `Dangl.Data.Shared.AspNetCore` package was updated to ASP.NET Core 2.2
+- Added `SpaStaticFileMiddlewareExtensions` for helper methods when working with Single Page Applications (SPAs)
 
 ## v1.0.10:
 - Add `CompressedRequestMiddleware` to support clients sending compressed request bodies